+++ kem/src/armv8/kyber_indcpa_armv8.c
@@ -27,6 +27,7 @@
 #include "kyber_indcpa_armv8.h"
 #include "kyber_poly_armv8.h"
 #include "kyber_polyvec_armv8.h"
+#include "kyber_kem_input_validation.h"
 #include "lc_sha3.h"
 #include "small_stack_support.h"
 #include "ret_checkers.h"
@@ -44,8 +45,8 @@
  * @param pk [in] pointer to the input public-key polyvec
  * @param seed [in] pointer to the input public seed
  */
-static void pack_pk(uint8_t r[LC_KYBER_INDCPA_PUBLICKEYBYTES], polyvec *pk,
-		    const uint8_t seed[LC_KYBER_SYMBYTES])
+static void pack_pk(uint8_t r[LC_KYBER_INDCPA_PUBLICKEYBYTES],
+		    const polyvec *pk, const uint8_t seed[LC_KYBER_SYMBYTES])
 {
 	polyvec_tobytes(r, pk);
 	memcpy(&r[LC_KYBER_POLYVECBYTES], seed, LC_KYBER_SYMBYTES);
@@ -72,7 +73,8 @@ static void unpack_pk(polyvec *pk, uint8_t seed[LC_KYBER_SYMBYTES],
  * @param r [out] pointer to output serialized secret key
  * @param sk [in] pointer to input vector of polynomials (secret key)
  */
-static void pack_sk(uint8_t r[LC_KYBER_INDCPA_SECRETKEYBYTES], polyvec *sk)
+static void pack_sk(uint8_t r[LC_KYBER_INDCPA_SECRETKEYBYTES],
+		    const polyvec *sk)
 {
 	polyvec_tobytes(r, sk);
 }
@@ -289,6 +291,7 @@ int indcpa_enc_armv8(uint8_t c[LC_KYBER_INDCPA_BYTES],
 	};
 	unsigned int i;
 	uint8_t nonce = 0, nonce2 = LC_KYBER_K;
+	int ret;
 	LC_DECLARE_MEM(ws, struct workspace, sizeof(uint64_t));
 
 	/*
@@ -297,6 +300,12 @@ int indcpa_enc_armv8(uint8_t c[LC_KYBER_INDCPA_BYTES],
 	 */
 	BUILD_BUG_ON(POLY_GETNOISE_ETA1_BUFSIZE < LC_KYBER_SYMBYTES);
 	unpack_pk(&ws->pkpv, ws->poly_getnoise_eta1_buf /* ws->seed */, pk);
+
+	/* Validate input */
+	CKINT(kyber_kem_iv_pk_modulus(pk, &ws->pkpv,
+				      ws->poly_getnoise_eta1_buf /* ws->seed */,
+				      pack_pk));
+
 	poly_frommsg(&ws->k, m);
 	gen_at(ws->at, ws->poly_getnoise_eta1_buf /* ws->seed */);
 
@@ -333,8 +342,9 @@ int indcpa_enc_armv8(uint8_t c[LC_KYBER_INDCPA_BYTES],
 
 	pack_ciphertext(c, &ws->b, &ws->v);
 
+out:
 	LC_RELEASE_MEM(ws);
-	return 0;
+	return ret;
 }
 
 int indcpa_dec_armv8(uint8_t m[LC_KYBER_INDCPA_MSGBYTES],
@@ -345,10 +355,14 @@ int indcpa_dec_armv8(uint8_t m[LC_KYBER_INDCPA_MSGBYTES],
 		polyvec b, skpv;
 		poly v, mp;
 	};
+	int ret;
 	LC_DECLARE_MEM(ws, struct workspace, sizeof(uint64_t));
 
 	unpack_ciphertext(&ws->b, &ws->v, c);
 	unpack_sk(&ws->skpv, sk);
+
+	/* Validate input */
+	CKINT(kyber_kem_iv_sk_modulus(sk, &ws->skpv, pack_sk));
 
 	polyvec_ntt(&ws->b);
 	polyvec_basemul_acc_montgomery(&ws->mp, &ws->skpv, &ws->b);
@@ -358,6 +372,7 @@ int indcpa_dec_armv8(uint8_t m[LC_KYBER_INDCPA_MSGBYTES],
 
 	poly_tomsg(m, &ws->mp);
 
+out:
 	LC_RELEASE_MEM(ws);
-	return 0;
+	return ret;
 }